fix(ci): resolve cloned_ref_to_slice_refs clippy lint (Nix cargo 1.93) (#407)
## What changed
Replace \&[x.clone()]\ with \std::slice::from_ref(&x)\ in three test
files to fix \clippy::cloned_ref_to_slice_refs\ lint on Nix CI (cargo
1.93.0).

## Files fixed
- \crates/tokmd-fun/tests/deep2.rs:413\
- \crates/tokmd-analysis-assets/tests/bdd.rs:779\
- \crates/tokmd-scan/tests/deep.rs:555\

## What ran
- CI truth (from #390 Nix Build failure)
- Local: verified pattern matches

## Determinism impact: none
## Contract impact: none
## Disposition: A) MERGE NOW

---------

+++ crates/tokmd-format/tests/snapshot_deep.rs
@@ -534,7 +534,7 @@ fn many_module_report() -> ModuleReport {
             module: format!("crates/mod_{:02}", i),
             code: (10 - i) * 100,
             lines: (10 - i) * 120,
-            files: (10 - i) as usize,
+            files: 10 - i,
             bytes: (10 - i) * 5000,
             tokens: (10 - i) * 250,
             avg_lines: 120,
